+++ backend/app/models/emotion_record.py
@@ -5,12 +5,12 @@ from app.config.database_config import db
 class EmotionRecord(db.Model):
     __tablename__ = 'emotion_record'
 
-    emotion_id = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
+    emotion_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     student_id = db.Column(db.String(20), db.ForeignKey('student_info.student_id', ondelete='SET NULL'), nullable=True)
     class_name = db.Column(db.String(50), nullable=True)
     source_type = db.Column(db.Enum('attendance', 'group_photo'), nullable=False)
-    attendance_record_id = db.Column(db.BigInteger, db.ForeignKey('attendance_record.record_id', ondelete='CASCADE'), nullable=True)
-    group_detail_id = db.Column(db.BigInteger, db.ForeignKey('group_photo_recognition_detail.detail_id', ondelete='CASCADE'), nullable=True)
+    attendance_record_id = db.Column(db.Integer, db.ForeignKey('attendance_record.record_id', ondelete='CASCADE'), nullable=True)
+    group_detail_id = db.Column(db.Integer, db.ForeignKey('group_photo_recognition_detail.detail_id', ondelete='CASCADE'), nullable=True)
     emotion = db.Column(db.String(20), nullable=False)
     confidence = db.Column(db.Numeric(5, 2), nullable=True)
     detected_at = db.Column(db.DateTime, nullable=False)
